% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taiga_nickname.R
\name{add_taiga_nickname}
\alias{add_taiga_nickname}
\title{Add a new taiga nickname}
\usage{
add_taiga_nickname(nickname, data_name, data_file, data_version = NA,
  filename = NULL)
}
\arguments{
\item{nickname:}{nickname for the taiga file}

\item{data_name:}{taiga data name}

\item{data_file:}{taiga data file}

\item{data_version:}{taiga data version}

\item{filename:}{name of the taiga nickname file if it exists}
}
\description{
Adds a new nickname to the taiga_nicknames file. Creates a new taiga_nicknames file if none found.
}
