% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_taiga_nickname.R
\name{call_taiga_nickname}
\alias{call_taiga_nickname}
\title{Load taiga data using a nickname}
\usage{
call_taiga_nickname(nickname, filename = NULL)
}
\arguments{
\item{nickname:}{nickname for taiga file you are referencing}

\item{filename:}{path to file of taiga nicknames if not just taiga_nicknames}
}
\description{
Loads data from taiga given most recent nickname pre-specified by the user.
}
