% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_ids.R
\name{change_ids}
\alias{change_ids}
\title{Map one set of identifiers to another}
\usage{
change_ids(input_df, map_df, from, to, dim)
}
\arguments{
\item{input_df:}{input data frame}

\item{map_df:}{input data frame of mapping ids}

\item{from:}{parameter in map_df containing the original ids}

\item{to:}{parameter in map_df containing the new ids}

\item{dim:}{whether you are changing rows (1) or columns (2)}
}
\description{
Maps identifiers given in `input_ids` to new identifiers contained in `map_df`
}
