% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cell_line_features_matrix.R
\name{create_cell_line_features_matrix}
\alias{create_cell_line_features_matrix}
\title{Prepare DepMap features data for upload to Taiga}
\usage{
create_cell_line_features_matrix()
}
\description{
Updates the cell line feature matrix as new data is added
}
