% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_annotations.R
\name{ensembl_hgnc_gene_pairing}
\alias{ensembl_hgnc_gene_pairing}
\title{Pair Ensembl gene ids with the correct HGNC gene symbol}
\usage{
ensembl_hgnc_gene_pairing(filename)
}
\arguments{
\item{filename:}{path to gencode gene annotations file, possible files are gencode.v19.annotation.gtf or gencode.v29.annotation.gff3}
}
\description{
Parses the GENCODE gene annotations file to get the Ensembl ID and HGNC symbols for genes
}
