% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gene_level_cn_for_taiga.R
\name{intersect_gene_with_copy_number}
\alias{intersect_gene_with_copy_number}
\title{Get a matrix of CN for genes}
\usage{
intersect_gene_with_copy_number(gene.gr, seg.gr, CN.column = "CN",
  gene.column = "Gene", do_parallel = F)
}
\arguments{
\item{gene.gr}{GRanges object of genes}

\item{seg.gr}{List of GRanges objects of CN segments}

\item{CN.column}{column to find CN data in seg.gr}

\item{gene.column}{gene.gr must contain a column with guide sequence}

\item{do_parallel}{run on multiple cores}
}
\value{
a matrix of CN data, each row is a gene, each column is a sample
}
\description{
Get a matrix of CN for genes
}
