% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_add_variant_annotations.R
\name{maf_add_variant_annotations}
\alias{maf_add_variant_annotations}
\title{Add mutation annotations}
\usage{
maf_add_variant_annotations(mut_maf)
}
\arguments{
\item{mut_maf:}{matrix MAF matrix}
}
\description{
Add columns to maf file describing each mutation as either a silent, damaging
non-conserving, or other-conserving mutations
}
