% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ssGSEA.R
\name{make_ssGSEA}
\alias{make_ssGSEA}
\title{Makes single sample GSEA for latest expression data}
\usage{
make_ssGSEA(data_mat, gmt_path, gsva_method = "ssgsea")
}
\arguments{
\item{data_mat:}{matrix of gene expression data where columns are HGNC symbol (Entrez ID) and rows are samples}

\item{gmt_path:}{path to gmt file of gene set object}

\item{gsva_method:}{method to use in estimation of gene-set enrichment scores per sample, default is ssgsea. 
ssGSEA calculates a gene set enrichment score per sample as the normalized difference in empirical cumulative distribution
functions of gene epxression ranks inside and outside the gene set}
}
\description{
Makes single sample GSEA for latest expression data
}
