% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dim_names.R
\name{map_dim_names}
\alias{map_dim_names}
\title{Map one set of row and column identifiers to another}
\usage{
map_dim_names(data_matrix, row_map, col_map, row_from, row_to, col_from,
  col_to)
}
\arguments{
\item{data_matrix:}{input data frame}

\item{row_map:}{data frame containing mapping for row identifiers}

\item{col_map:}{data frame containing mapping for column identifiers}

\item{row_from:}{parameter in row_map containing the original row identifiers}

\item{row_to:}{parameter in row_map contaning the new row identifiers}

\item{col_from:}{parameter in col_map containing the original column identifiers}

\item{col_to:}{parameter in col_map containing the new column identifiers}
}
\description{
Maps matrix row and column names to new identifiers given in `row_map` and `col_map`
}
