% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_maf_to_binary_matrix.R
\name{mutation_maf_to_binary_matrix}
\alias{mutation_maf_to_binary_matrix}
\title{Prepare DepMap mutation data for upload to Taiga}
\usage{
mutation_maf_to_binary_matrix(mut_maf, gene_symbol_col = "Hugo_Symbol",
  gene_id_col = "Entrez_Gene_Id", damaging = F, other = F,
  hotspot = F)
}
\arguments{
\item{mut_maf:}{maf file}

\item{gene_symbol_col:}{name of column containing the gene symbols}

\item{gene_id_col:}{name of column containing the gene ids}

\item{damaging:}{initially is set to F, set to T to create binary matrix of damaging mutations}

\item{other:}{initially is set to F, set to T to create binary matrix of conserving and non conserving mutations}

\item{hotspot:}{initially is set to F, set to T to create binary matrix of hotspot mutations}
}
\description{
Assumes that mutation calls have been provided for each cell line
}
