% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_depmap_fusion_data_for_taiga.R
\name{prepare_depmap_fusion_data_for_taiga}
\alias{prepare_depmap_fusion_data_for_taiga}
\title{Prepare DepMap fusion data for upload to Taiga}
\usage{
prepare_depmap_fusion_data_for_taiga(fusion_data)
}
\arguments{
\item{fusion_data:}{fusion data matrix}
}
\description{
Adjust column names and syntax of gene names for fusion data
}
