% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_depmap_rppa_for_taiga.R
\name{prepare_depmap_rppa_for_taiga}
\alias{prepare_depmap_rppa_for_taiga}
\title{Prepare DepMap RPPA data for upload to Taiga}
\usage{
prepare_depmap_rppa_for_taiga(rppa_filename, antibody_info_filename,
  antibody_id_col = "Antibody_Name", gene_id_col = "Target_Genes")
}
\arguments{
\item{rppa_filename:}{path to rppa file}

\item{antibody_info_filename:}{path to antibody info file}

\item{antibody_id_col:}{name of column containing the antibody names}

\item{gene_id_col:}{name of column containing the gene names}
}
\description{
Assumes that RPPA data arrive as a .csv file with CCLE cell line
names on the rows and antibody identifiers on the columns. The antibody info file
is assumed to be a .csv file with a mapping from antibody identifier to HUGO symbol.
}
