% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_depmap_rnaseq_for_taiga.R
\name{prepare_depmap_TPM_for_taiga}
\alias{prepare_depmap_TPM_for_taiga}
\title{Prepare DepMap TPM RNAseq data for upload to Taiga}
\usage{
prepare_depmap_TPM_for_taiga(rnaseq_data, log_transform = F,
  just_protein_coding = F,
  gencode_file = "~/data_files/gene_annotations/gencode.v19.annotation.gtf")
}
\arguments{
\item{rnaseq_data:}{rnaseq file, assumes rows are genes, gene_id and transcript_id columns, and rest of columns are samples}

\item{log_transform:}{whether or not to log2 transform the data}

\item{just_protein_coding:}{whether or not to just subset to protein coding genes and use entrez ids}

\item{gencode_file:}{path to the gencode file with Ensembl ID and HGNC symbol pairings}
}
\description{
Takes in a matrix of TPM values then log2 + 1 transforms it and returns gene expression matrix of
cell lines by genes
}
