% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_depmap_rnaseq_for_taiga.R
\name{prepare_depmap_transcripts_for_taiga}
\alias{prepare_depmap_transcripts_for_taiga}
\title{Prepare DepMap TPM transcript RNAseq data for upload to Taiga}
\usage{
prepare_depmap_transcripts_for_taiga(rnaseq_data,
  gencode_file = "~/data_files/gene_annotations/gencode.v19.annotation.gtf")
}
\arguments{
\item{rnaseq_data:}{rnaseq file, assumes rows are genes, gene_id and transcript_id columns, and rest of columns are samples}

\item{gencode_file:}{path to the gencode file with Ensembl ID and HGNC symbol pairings}
}
\description{
Takes in a matrix of transcript data and returns transcript data with gene names
}
