% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gene_level_cn_for_taiga.R
\name{prepare_gene_level_cn_for_taiga}
\alias{prepare_gene_level_cn_for_taiga}
\title{Prepare gene-level copy number matrix for upload to taiga}
\usage{
prepare_gene_level_cn_for_taiga(filename, ccds_filename, do_parallel)
}
\arguments{
\item{filename:}{path copy number file}

\item{ccds_filename:}{path to ccds file}

\item{do_parallel:}{whether or not to run in parallel}
}
\description{
Assumes input filename is path to a segmented copy number file
}
