% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_miRNA_for_taiga.R
\name{prepare_miRNA_for_taiga}
\alias{prepare_miRNA_for_taiga}
\title{Prepare miRNA data for upload to Taiga}
\usage{
prepare_miRNA_for_taiga(filename, log_transform = T)
}
\arguments{
\item{filename:}{path to miRNA file}

\item{log_transform:}{boolean, whether or not log2(.+50) transform the data}
}
\description{
Assumes that miRNA data arrive in .gct file (tab-delimited) 
with `Name`, `Description`, and CCLE cell line names as columns. The `Name` field contains
the Accession number for each gene, and the `Description` field 
contains the miRNA ID.
}
