% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_RRBS_for_taiga.R
\name{prepare_RRBS_for_taiga}
\alias{prepare_RRBS_for_taiga}
\title{Prepare RRBS data for upload to Taiga}
\usage{
prepare_RRBS_for_taiga(filename1, filename2)
}
\arguments{
\item{filename1:}{path to RRBS_1kb file}

\item{filename2:}{path to RRBS_CpG file}
}
\description{
Assumes that RRBS data arrive in two .txt files (tab-delimited) 
The first is 1kb data and the other is CpG cluster data
The 1kb dataset has columns TSS_id, gene (which is HGNC symbol), chr, fpos, tpos, strand, avg_coverage, and the CCLE IDs
The CpG cluster dataset has columns cluster_id, gene_name (which is HGNC symbol), RefSeq_id, CpG_sites_hg19, avg_coverage and the CCLE IDs                         "avg_coverage"
Return four datasets, methylation data with columns of genes (HGNC symbols) and rows of Broad cell line IDs for 1kb data and CpG data
and meta data file for the 1kb data and CpG data
}
