% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_annotations.R
\name{protein_coding_genes}
\alias{protein_coding_genes}
\title{Identify the gene names for protein coding genes}
\usage{
protein_coding_genes(gene_map, hgnc_gene_dataset)
}
\arguments{
\item{gene_map:}{matrix containing the Ensembl ID and HGNC symbol pairings, the Ensembl ID column must be named ensembl_gene_id}

\item{hgnc_gene_dataset:}{dataset from HGNC with gene annotations}
}
\description{
gets the Ensembl ID, HGNC symbol, and Entrez ID for protein coding genes
}
