% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_gene_level_cn.R
\name{subset_gene_level_cn}
\alias{subset_gene_level_cn}
\title{Replace cell lines with embargoed Broad WES CN data with chordoma, Sanger or SNP CN data}
\usage{
subset_gene_level_cn(gene_cn_broad, gene_cn_chordoma, gene_cn_sanger,
  gene_cn_SNP, replace_CLs)
}
\arguments{
\item{gene_cn_broad:}{gene level copy number data containing data from internal and blacklisted cell lines}

\item{gene_cn_chordoma:}{gene level copy number data for the chordoma lines}

\item{gene_cn_sanger:}{gene level copy number data from Sanger}

\item{gene_cn_SNP:}{gene level copy number data using SNP data}

\item{replace_CLs:}{cell lines that are not allowed for this dataset (have embargoed WES)}
}
\description{

}
