% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_annotations.R
\name{test_gencode_ids}
\alias{test_gencode_ids}
\title{Test that GENCODE Ensembl ID-HGNC symbol pairings match HGNC dataswt}
\usage{
test_gencode_ids(gene_map, hgnc_gene_dataset)
}
\arguments{
\item{gene_map:}{matrix containing the Ensembl ID and HGNC symbol pairings from GENCODE}

\item{version:}{dataset from HGNC with gene annotations}
}
\description{
confirms that gene ID pairings are correct
}
