% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_new_data.R
\name{test_new_data}
\alias{test_new_data}
\title{Test the new data to make sure it mostly matches the previous dataset}
\usage{
test_new_data(new_data, previous_data, data_type = "RNAseq")
}
\arguments{
\item{new_data:}{newest version of the RNAseq dataset}

\item{previous_data:}{previous release of RNAseq data}

\item{data_type:}{either RNAseq data or CN for copy number data}
}
\description{
Look at the correlation between cell lines in common between the new data and the previous release of data
}
